+++ hesperidescli/applications/platforms/snapshots/snapshots.py
@@ -1,75 +1,24 @@
+# pylint: disable=unused-argument
 import click
 
-from hesperidescli import utils
-from hesperidescli.client import Client
 
-
-@click.command("get-application-platform-snapshots")
+@click.command("get-application-platform-snapshots", deprecated=True)
 @click.option("--application-name")
 @click.option("--platform-name")
 def get_application_platform_snapshots(application_name, platform_name):
-    if application_name is None:
-        print("--application-name required")
-        raise click.Abort()
-    if platform_name is None:
-        print("--platform-name required")
-        raise click.Abort()
-    client = Client()
-    response = client.get(
-        "/rest/applications/"
-        + application_name
-        + "/platforms/"
-        + platform_name
-        + "/snapshots"
-    )
-    utils.pretty_print(response)
+    return
 
 
-@click.command("restore-application-platform-snapshots")
+@click.command("restore-application-platform-snapshots", deprecated=True)
 @click.option("--application-name")
 @click.option("--platform-name")
 def restore_application_platform_snapshots(application_name, platform_name):
-    if application_name is None:
-        print("--application-name required")
-        raise click.Abort()
-    if platform_name is None:
-        print("--platform-name required")
-        raise click.Abort()
-    client = Client()
-    response = client.post(
-        "/rest/applications/"
-        + application_name
-        + "/platforms/"
-        + platform_name
-        + "/restore_snapshot"
-    )
-    utils.pretty_print(response)
+    return
 
 
-@click.command("take-application-platform-snapshot")
+@click.command("take-application-platform-snapshot", deprecated=True)
 @click.option("--application-name")
 @click.option("--platform-name")
 @click.option("--body")
 def take_application_platform_snapshot(application_name, platform_name, body):
-    if application_name is None:
-        print("--application-name required")
-        raise click.Abort()
-    if platform_name is None:
-        print("--platform-name required")
-        raise click.Abort()
-    if body is None:
-        print("--body required")
-        raise click.Abort()
-    file = open(body, "r")
-    file_body = file.read()
-    file.close()
-    client = Client()
-    response = client.post(
-        "/rest/applications/"
-        + application_name
-        + "/platforms/"
-        + platform_name
-        + "/take_snapshot",
-        body=file_body,
-    )
-    utils.pretty_print(response)
+    return