+++ hesperidescli/templates/templates.py
@@ -5,59 +5,42 @@ from hesperidescli.client import Client
 
 
 @click.command("create-template-package")
+@click.argument("body")
 @click.option("--from-package-name")
 @click.option("--from-package-version")
 @click.option("--from-is-working-copy", is_flag=True)
-@click.option("--body")
 def create_template_package(
-    from_package_name, from_package_version, from_is_working_copy, body
+    body, from_package_name, from_package_version, from_is_working_copy
 ):
     params = {}
     if from_package_name is None and from_package_version:
-        print("--from-package-name required when --from-package-version is given")
-        raise click.Abort()
+        raise click.BadParameter(
+            "--from-package-name required when --from-package-version is given"
+        )
     if from_package_name and from_package_version is None:
-        print("--from-package-version required when --from-package-name is given")
-        raise click.Abort()
-    if body is None:
-        print("--body required")
-        raise click.Abort()
+        raise click.BadParameter(
+            "--from-package-version required when --from-package-name is given"
+        )
     params["from_package_name"] = from_package_name
     params["from_package_version"] = from_package_version
     params["from_is_working_copy"] = from_is_working_copy
-    file = open(body, "r")
-    file_body = file.read()
-    file.close()
-    client = Client()
-    response = client.post("/rest/templates/packages", params=params, body=file_body)
+    with open(body, "r") as body_file:
+        file_body = body_file.read()
+    response = Client().post("/rest/templates/packages", params=params, body=file_body)
     utils.pretty_print(response)
 
 
 @click.command("create-template-package-workingcopy")
-@click.option("--package-name")
-@click.option("--package-version")
-@click.option("--template-name")
-@click.option("--body")
+@click.argument("body")
+@click.option("--package-name", required=True)
+@click.option("--package-version", required=True)
+@click.option("--template-name", required=True)
 def create_template_package_workingcopy(
-    package_name, package_version, template_name, body
+    body, package_name, package_version, template_name
 ):
-    if package_name is None:
-        print("--package-name required")
-        raise click.Abort()
-    if package_version is None:
-        print("--package-version required")
-        raise click.Abort()
-    if template_name is None:
-        print("--template-name required")
-        raise click.Abort()
-    if body is None:
-        print("--body required")
-        raise click.Abort()
-    file = open(body, "r")
-    file_body = file.read()
-    file.close()
-    client = Client()
-    response = client.post(
+    with open(body, "r") as body_file:
+        file_body = body_file.read()
+    response = Client().post(
         "/rest/templates/packages/"
         + package_name
         + "/"
@@ -70,34 +53,20 @@ def create_template_package_workingcopy(
 
 
 @click.command("delete-template-package-release")
-@click.option("--package-name")
-@click.option("--package-version")
+@click.option("--package-name", required=True)
+@click.option("--package-version", required=True)
 def delete_template_package_release(package_name, package_version):
-    if package_name is None:
-        print("--package-name required")
-        raise click.Abort()
-    if package_version is None:
-        print("--package-version required")
-        raise click.Abort()
-    client = Client()
-    response = client.delete(
+    response = Client().delete(
         "/rest/templates/packages/" + package_name + "/" + package_version + "/release"
     )
     utils.pretty_print(response)
 
 
 @click.command("delete-template-package-workingcopy")
-@click.option("--package-name")
-@click.option("--package-version")
+@click.option("--package-name", required=True)
+@click.option("--package-version", required=True)
 def delete_template_package_workingcopy(package_name, package_version):
-    if package_name is None:
-        print("--package-name required")
-        raise click.Abort()
-    if package_version is None:
-        print("--package-version required")
-        raise click.Abort()
-    client = Client()
-    response = client.delete(
+    response = Client().delete(
         "/rest/templates/packages/"
         + package_name
         + "/"
@@ -108,21 +77,11 @@ def delete_template_package_workingcopy(package_name, package_version):
 
 
 @click.command("get-template-package-release")
-@click.option("--package-name")
-@click.option("--package-version")
-@click.option("--template-name")
+@click.option("--package-name", required=True)
+@click.option("--package-version", required=True)
+@click.option("--template-name", required=True)
 def get_template_package_release(package_name, package_version, template_name):
-    if package_name is None:
-        print("--package-name required")
-        raise click.Abort()
-    if package_version is None:
-        print("--package-version required")
-        raise click.Abort()
-    if template_name is None:
-        print("--template-name required")
-        raise click.Abort()
-    client = Client()
-    response = client.get(
+    response = Client().get(
         "/rest/templates/packages/"
         + package_name
         + "/"
@@ -134,17 +93,10 @@ def get_template_package_release(package_name, package_version, template_name):
 
 
 @click.command("get-template-package-release-model")
-@click.option("--package-name")
-@click.option("--package-version")
+@click.option("--package-name", required=True)
+@click.option("--package-version", required=True)
 def get_template_package_release_model(package_name, package_version):
-    if package_name is None:
-        print("--package-name required")
-        raise click.Abort()
-    if package_version is None:
-        print("--package-version required")
-        raise click.Abort()
-    client = Client()
-    response = client.get(
+    response = Client().get(
         "/rest/templates/packages/"
         + package_name
         + "/"
@@ -155,21 +107,11 @@ def get_template_package_release_model(package_name, package_version):
 
 
 @click.command("get-template-package-workingcopy")
-@click.option("--package-name")
-@click.option("--package-version")
-@click.option("--template-name")
+@click.option("--package-name", required=True)
+@click.option("--package-version", required=True)
+@click.option("--template-name", required=True)
 def get_template_package_workingcopy(package_name, package_version, template_name):
-    if package_name is None:
-        print("--package-name required")
-        raise click.Abort()
-    if package_version is None:
-        print("--package-version required")
-        raise click.Abort()
-    if template_name is None:
-        print("--template-name required")
-        raise click.Abort()
-    client = Client()
-    response = client.get(
+    response = Client().get(
         "/rest/templates/packages/"
         + package_name
         + "/"
@@ -181,17 +123,10 @@ def get_template_package_workingcopy(package_name, package_version, template_nam
 
 
 @click.command("get-template-package-workingcopy-model")
-@click.option("--package-name")
-@click.option("--package-version")
+@click.option("--package-name", required=True)
+@click.option("--package-version", required=True)
 def get_template_package_workingcopy_model(package_name, package_version):
-    if package_name is None:
-        print("--package-name required")
-        raise click.Abort()
-    if package_version is None:
-        print("--package-version required")
-        raise click.Abort()
-    client = Client()
-    response = client.get(
+    response = Client().get(
         "/rest/templates/packages/"
         + package_name
         + "/"
@@ -202,17 +137,10 @@ def get_template_package_workingcopy_model(package_name, package_version):
 
 
 @click.command("get-templates-packages-release")
-@click.option("--package-name")
-@click.option("--package-version")
+@click.option("--package-name", required=True)
+@click.option("--package-version", required=True)
 def get_templates_packages_release(package_name, package_version):
-    if package_name is None:
-        print("--package-name required")
-        raise click.Abort()
-    if package_version is None:
-        print("--package-version required")
-        raise click.Abort()
-    client = Client()
-    response = client.get(
+    response = Client().get(
         "/rest/templates/packages/"
         + package_name
         + "/"
@@ -223,17 +151,10 @@ def get_templates_packages_release(package_name, package_version):
 
 
 @click.command("get-templates-packages-workingcopy")
-@click.option("--package-name")
-@click.option("--package-version")
+@click.option("--package-name", required=True)
+@click.option("--package-version", required=True)
 def get_templates_packages_workingcopy(package_name, package_version):
-    if package_name is None:
-        print("--package-name required")
-        raise click.Abort()
-    if package_version is None:
-        print("--package-version required")
-        raise click.Abort()
-    client = Client()
-    response = client.get(
+    response = Client().get(
         "/rest/templates/packages/"
         + package_name
         + "/"
@@ -244,43 +165,25 @@ def get_templates_packages_workingcopy(package_name, package_version):
 
 
 @click.command("perform-search-templates-packages")
-@click.option("--terms")
+@click.argument("terms")
 def perform_search_templates_packages(terms):
-    params = {}
-    if terms is None:
-        print("--terms required")
-        raise click.Abort()
-    params["terms"] = terms
-    client = Client()
-    response = client.post("/rest/templates/packages/perform_search", params=params)
+    response = Client().post(
+        "/rest/templates/packages/perform_search", params={"terms": terms}
+    )
     utils.pretty_print(response)
 
 
 @click.command("update-template-package-workingcopy")
-@click.option("--package-name")
-@click.option("--package-version")
-@click.option("--template-name")
-@click.option("--body")
+@click.argument("body")
+@click.option("--package-name", required=True)
+@click.option("--package-version", required=True)
+@click.option("--template-name", required=True)
 def update_template_package_workingcopy(
-    package_name, package_version, template_name, body
+    body, package_name, package_version, template_name
 ):
-    if package_name is None:
-        print("--package-name required")
-        raise click.Abort()
-    if package_version is None:
-        print("--package-version required")
-        raise click.Abort()
-    if template_name is None:
-        print("--template-name required")
-        raise click.Abort()
-    if body is None:
-        print("--body required")
-        raise click.Abort()
-    file = open(body, "r")
-    file_body = file.read()
-    file.close()
-    client = Client()
-    response = client.put(
+    with open(body, "r") as body_file:
+        file_body = body_file.read()
+    response = Client().put(
         "/rest/templates/packages/"
         + package_name
         + "/"